pyproject.toml
[project]
name = "nukitashi-cgtools"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "lupa>=2.5",
    "pillow>=11.3.0",
]
